feat(calculators): merge rent estimate and ceiling into rent analyser
Combines RentEstimate and RentCeilingCalculator into a unified
RentAnalyser component. Shared state (postcode, sizeSqm, buildingYear)
is lifted to the parent and pre-populates Section 2 from Section 1.
Old tab values rent-estimate and rent-ceiling fall through to the new
rent-analyser tab for graceful URL handling.
