+++ frontend/src/components/Calculators/RentAnalyser/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Rent Analyser index
+ */
+
+export { RentAnalyser } from "./RentAnalyser"
