+++ frontend/src/components/Calculators/index.ts
@@ -12,8 +12,7 @@ export { MortgageAmortisation } from "./MortgageAmortisation"
 export { MortgageEligibilityChecker } from "./MortgageEligibilityChecker"
 export { OwnershipComparison } from "./OwnershipComparison"
 export { PropertyEvaluationCalculator } from "./PropertyEvaluationCalculator"
-export { RentCeilingCalculator } from "./RentCeilingCalculator"
-export { RentEstimate } from "./RentEstimate"
+export { RentAnalyser } from "./RentAnalyser"
 export { ROICalculator } from "./ROICalculator"
 export { SpeculationTaxCalculator } from "./SpeculationTaxCalculator"
 export { StateComparison } from "./StateComparison"
